+++ src/pages/components/DateSelector.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { DateSelector } from './DateSelector'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 const meta: Meta<typeof DateSelector> = {
   title: 'Components/DateSelector',
