+++ src/pages/components/DateTimePicker.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useState } from 'react'
-import { Dayjs } from 'src/pages/components/utils/dayjs'
+import { Dayjs } from 'src/dayjs'
 
 export function useDate(initialValue: Dayjs) {
   const [date, setDate] = useState<Dayjs>(initialValue)
