+++ src/pages/components/timeline/TimelineBoard.tsx
@@ -10,7 +10,7 @@ import { Coordinates } from 'src/model/location'
 import { HorizontalLine } from 'src/pages/components/timeline/HorizontalLine'
 import { Timeline } from 'src/pages/components/timeline/Timeline'
 import { PointType } from 'src/pages/components/timeline/TimelinePoint'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 const COLUMN_WIDTH = 140
 export const PADDING = 10
