+++ src/pages/dashboard/ArrivalByTimeChart/ArrivalByTimeChart.tsx
@@ -8,7 +8,7 @@ import {
   XAxis,
   YAxis,
 } from 'recharts'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 import './ArrivalByTimeChats.scss'
 
 /**
