+++ src/pages/dashboard/WorstLinesChart/WorstLinesChart.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import WorstLinesChart from './WorstLinesChart'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 const meta: Meta<typeof WorstLinesChart> = {
   component: WorstLinesChart,
