+++ src/pages/lineProfile/LineProfileRide.tsx
@@ -3,7 +3,7 @@ import { InfoItem, InfoTable } from '../components/InfoTable'
 import { vehicleIDFormat } from '../components/utils/rotueUtils'
 import { VehicleLocation } from 'src/model/vehicleLocation'
 import Widget from 'src/shared/Widget'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 export const LineProfileRide = ({ point }: { point?: VehicleLocation }) => {
   const { t } = useTranslation()
