+++ src/pages/operator/OperatorGaps.tsx
@@ -6,7 +6,7 @@ import { Cell, Pie, PieChart } from 'recharts'
 import { InfoItem, InfoTable } from '../components/InfoTable'
 import { useGroupBy } from 'src/api/groupByService'
 import Widget from 'src/shared/Widget'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 export const OperatorGaps = ({
   operatorId,
